change int() to float() in NIST energy method

+++ src/rxn_network/entries/nist.py
@@ -72,7 +72,7 @@ class NISTReferenceEntry(Entry):
         """
         data = NISTReferenceEntry.REFERENCES[formula]
         if temperature % 100 > 0:
-            g_interp = interp1d([int(t) for t in data.keys()], list(data.values()))
+            g_interp = interp1d([float(t) for t in data.keys()], list(data.values()))
             return g_interp(temperature)
 
         return data[str(temperature)]
